add definition in SubLayer.h, fix some bug.

+++ Convolute.h
@@ -1,4 +1,4 @@
 #include <opencv/cv.h>
 using namespace cv;
 
-void Convolute(Mat* inputArray, Mat* outputArray, int GaussianVariant);
+void Convolute(Mat* inputArray, Mat* outputArray, double GaussianVariant);
